can_autobaud: Don't use baudrate from boot msg
Presumably, the deployed bootloaders have the bug where baudrate
transmitted in the boot msg is not correct.

For Pixhawk 6X the timing is just right somehow to get it stuck at
baudrate 500000 instead of 1000000.

Disabling reading from boot msg seems fixes this.

The correct fix would probably be in the bootloader but it's not
very realistic to update the bootloaders in the wild.

Signed-off-by: Julian Oes <julian@oes.ch>